Rename identifiers of each projects and add comments
