+++ vertx-gaia/vertx-co/pom.xml
@@ -10,7 +10,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>vertx-co</artifactId>
-    <name>( Vertx Core ) Zero Core Infrastructure</name>
+    <name>( Vertx Core ) vertx-co: Zero Infrastructure</name>
     <dependencies>
         <dependency>
             <groupId>io.vertx</groupId>
