+++ vertx-gaia/vertx-co/src/main/java/io/vertx/up/annotations/Phase.java
@@ -18,7 +18,7 @@ public @interface Phase {
     /**
      * Phase value
      *
-     * @return
+     * @return Authentication phase for custom workflow
      */
     AuthPhase value();
 }
