+++ vertx-gaia/vertx-co/src/main/java/io/vertx/zero/config/NodeVisitor.java
@@ -23,7 +23,7 @@ public interface NodeVisitor
     /**
      * Get cluster configuration from vertx initialization.
      *
-     * @return
+     * @return io.vertx.core.ClusterOptions that defined by zero.
      */
     ClusterOptions getCluster();
 }
