+++ vertx-gaia/vertx-co/src/main/java/io/vertx/zero/marshal/Visitor.java
@@ -4,8 +4,8 @@ import io.vertx.zero.exception.ZeroException;
 
 public interface Visitor<T> {
     /**
-     * @return
-     * @throws ZeroException
+     * @return The generic type of config
+     * @throws ZeroException The zero exception that prevent start up
      */
     T visit(String... keys) throws ZeroException;
 }
