+++ vertx-gaia/vertx-ifx/zero-ali-sms/pom.xml
@@ -8,6 +8,6 @@
         <version>0.4.8-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>vertx-ali-sms</artifactId>
-    <name>( Vertx Infix ) Ali Yun SMS Service</name>
+    <artifactId>zero-ali-sms</artifactId>
+    <name>( Vertx Infix ) zero-ali-sms: Ali Yun SMS</name>
 </project>
