+++ vertx-gaia/vertx-ifx/zero-ali-sms/src/main/java/io/vertx/tp/ali/sms/AliSmsClient.java
@@ -1,0 +1,14 @@
+package io.vertx.tp.ali.sms;
+
+import io.vertx.codegen.annotations.Fluent;
+import io.vertx.core.json.JsonObject;
+
+/**
+ * AliSmsClient for platform of https://dysms.console.aliyun.com/dysms.htm
+ * Message open sdk
+ */
+public interface AliSmsClient {
+    
+    @Fluent
+    AliSmsClient init(JsonObject params);
+}
