+++ vertx-gaia/vertx-ifx/zero-ali-sms/src/main/java/io/vertx/tp/ali/sms/AliSmsCodes.java
@@ -1,0 +1,4 @@
+package io.vertx.tp.ali.sms;
+
+public interface AliSmsCodes {
+}
