+++ vertx-gaia/vertx-ifx/zero-ali-sms/src/main/resources/ke/rules/ali-sms.yml
@@ -1,0 +1,6 @@
+required:
+- access_id
+- access_secret
+typed:
+  access_id: STRING
+  access_secret: STRING
