+++ vertx-gaia/vertx-ifx/zero-ali-sms/src/test/resources/vertx-tp.yml
@@ -1,0 +1,5 @@
+ali:
+  sms:
+    config:
+      access_id: LTAIGXqTuTkPPTm6
+      access_secret: lHyKexT3i7e613gHX6k4NRyatZa3y0
