+++ vertx-gaia/vertx-rx/pom.xml
@@ -8,7 +8,7 @@
         <version>0.4.8-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <name>( Vertx Core ) Zero Up RxJava Container</name>
+    <name>( Vertx Core ) vertx-rx: Zero Up RxJava Container</name>
     <artifactId>vertx-rx</artifactId>
     <dependencies>
         <dependency>
