+++ vertx-gaia/vertx-up/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>vertx-up</artifactId>
-    <name>( Vertx Core ) Zero Up Web Container</name>
+    <name>( Vertx Core ) vertx-up: Zero Up Web Container</name>
     <dependencies>
         <!-- Service Discovery -->
         <dependency>
